Better downloader structure
