+++ .gitignore
@@ -11,3 +11,4 @@ target/
 sources.tmp
 Apollo.jar
 Apolloservice.jar
+*.iml
