+++ src/main/java/com/apollocurrency/aplwallet/apl/updater/downloader/DownloadExecutor.java
@@ -1,0 +1,8 @@
+package com.apollocurrency.aplwallet.apl.updater.downloader;
+
+import java.io.IOException;
+import java.nio.file.Path;
+
+public interface DownloadExecutor {
+    Path download(String uri) throws IOException;
+}
